!578 Reconstruct the time processing system of X-Kernel
* chore: drop committed build artifacts
* fix(time): make wall-clock updates atomic and range-checked
* feat: add unittest dependency and update test configuration
* refactor: split ktime into value types and timekeeping layer
* fix: drop leftover wall_time reference in sys_time unittest
* chore: drop orphaned rsext4_vfs bridge
* fix: integrate wall-clock setters with ktime time model
* fix: gate device tree RTC discovery by feature
* feat: enhance timer functionality with deadline validation and state p…
* fix: address time handling review feedback
* refactor: replace wall_time with realtime in various components
* feat: add posix-types for CPU-time clock ticks and refactor time handling
* Refactor time-related structures and improve time handling
* feat: integrate ktime dependency and refactor time functions
* Refactor time handling in TEE kernel

+++ io/watchdog/src/watchdog_task.rs
@@ -60,6 +60,6 @@ impl WatchdogTask for MutexDeadlockCheck {
     }
 
     fn check(&self) -> bool {
-        ktask::check_mutex_deadlock(khal::time::now_ticks() as usize)
+        ktask::check_mutex_deadlock(khal::time::now_ticks())
     }
 }
