+++ task/ksync/src/mutex.rs
@@ -171,7 +171,7 @@ unsafe impl lock_api::RawMutex for RawMutex {
             #[cfg(feature = "watchdog")]
             current()
                 .inner()
-                .set_waiting_lock(self as *const _ as usize, khal::time::now_ticks() as usize);
+                .set_waiting_lock(self as *const _ as usize, khal::time::now_ticks());
             block_on(listener);
             owner_id = self.owner_id.load(Ordering::Acquire);
         }
